Restructure simulator
